% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGA_Tour_2021_Stats.R
\docType{data}
\name{PGA_Tour_2021_Stats}
\alias{PGA_Tour_2021_Stats}
\title{PGA Tour 2021 Stats}
\format{
A data frame of 37044 rows and 5 columns
\describe{
\item{Player_Name}{Character value containing names of players}
\item{Category}{Character value containing five major categories: Putting, Approach the Green, Around the Green, Off the Tee and Scoring}
\item{Sub_Category}{Character value containing different sub categories for each major category}
\item{Measure_Type}{Character value containing different measure types for each category/sub_category. This can be Number of Rounds played, Average Distance and so on}
\item{Value}{Numeric value according to each Measure_Type}
}
}
\source{
\url{https://www.kaggle.com/lilcobeeee/2021-pga-tour-stats/}
}
\usage{
PGA_Tour_2021_Stats
}
\description{
A dataset containing stats from the 2021 Season sourced from Kaggle
}
\keyword{datasets}
